non 110 support and fixed spec change issue
added ratings tables containing base amounts for all levels, and an associated level check to pull the correct base amount for your current level.

Also moved function that was pulling the current mastery coef into the mouseover call. This was done to allow for the correct mastery coef to be used at all times. Old method required a UI reload to update the coef if the user changed specs.