Fixed lua nil error after first login with any character
Fixed a lua nil error which occour after the addon is loaded. To avoid the error we have to register our hook after everything is loaded and all item infos are available. To do this we register the "PLAYER_ENTERING_WORLD" event and create the hook inside the event function